feature: product detail component
