+++ frontend/src/components/index.tsx
@@ -3,11 +3,15 @@ import {SearchBoxComponent} from './search-box/search-box.component';
 import {BreadcrumbComponent} from './breadcrumb/breadcrumb.component';
 import {ItemListComponent} from './item-list/item-list.component';
 import {AccentButton} from './accent-button/accent-button.component';
+import {ProductDetail} from './product-detail/product-detail.component';
+import {Price} from './price/price.component';
 
 export {
   SearchBarComponent, 
   SearchBoxComponent, 
   BreadcrumbComponent, 
   ItemListComponent,
-  AccentButton
+  AccentButton,
+  ProductDetail,
+  Price
 }
