+++ frontend/src/components/price/price.component.jsx
@@ -1,0 +1,20 @@
+import NumberFormat from 'react-number-format';
+import "./price.component.scss";
+
+function Price(props) {
+  const price = props.price;
+
+  return (
+    <NumberFormat
+      className="price-value"
+      value={price.amount}
+      displayType={'text'}
+      thousandSeparator={'.'}
+      decimalSeparator={','}
+      decimalScale={price.decimals}
+      prefix={'$ '}
+    />
+  )
+}
+
+export { Price }
