+++ frontend/src/components/price/price.component.scss
@@ -1,0 +1,3 @@
+.price-value {
+  white-space: nowrap;
+}
