+++ frontend/src/components/product-detail/product-detail.component.jsx
@@ -1,0 +1,42 @@
+
+import {
+  AccentButton,
+  Price
+} from '../../components';
+import "./product-detail.component.scss";
+
+function ProductDetail(props) {
+  const item = props.item;
+  const condition = item.condition === 'new' ? 'Nuevo' : 'Usado';
+  const condition_sold = condition.concat(' - ').concat(item.sold_quantity).concat(' vendidos');
+
+  function comprarAction() {
+    alert('Compraste el producto!!');
+  }
+
+  return (
+    <div className="product">
+      <div className="product-info">
+        <div className="picture">
+          <img
+            src={item.picture}
+            alt={item.title} />
+        </div>
+        <div className="panel">
+          <div className="condition-sold">{condition_sold}</div>
+          <div className="item-title">{item.title}</div>
+          <div className="price">
+            <Price price={item.price} />
+          </div>
+          <AccentButton onClick={comprarAction}>Comprar</AccentButton>
+        </div>
+      </div>
+      <div className="description">
+        <div className="title">Descripción del producto</div>
+        <div className="text">{item.description}</div>
+      </div>
+    </div>
+  )
+}
+
+export { ProductDetail };
