+++ frontend/src/components/product-detail/product-detail.component.scss
@@ -1,0 +1,61 @@
+@import 'colors';
+
+.product {
+  display: flex;
+  flex-direction: column;
+
+  .product-info {  
+    display: flex;
+    flex-wrap: wrap;
+    margin-bottom: 120px;
+    
+    .picture {
+      width: 680px;
+      max-height: 680px;
+      border-radius: 4px;
+      overflow: hidden;
+      display: flex;
+      flex: 0 1 680px;
+      justify-content: center;
+
+      img {
+        width: 100%;
+        object-fit: contain;
+      }
+    }
+
+    .panel {
+      display: flex;
+      flex: 1 1 200px;
+      flex-direction: column;
+      max-width: 300px;
+      padding-left: 32px;
+      padding-top: 32px;
+
+      .condition-sold {
+        font-size: 14px;
+        margin-bottom: 16px;
+      }
+
+      .item-title {
+        font-size: 24px;
+        margin-bottom: 32px;
+      }
+
+      .price {
+        font-size: 46px;
+        margin-bottom: 32px;
+      }
+    }
+  }
+
+  .description {
+    .title {
+      font-size: 28px;
+      margin-bottom: 32px;
+    }
+    .text {
+      font-size: 16px;
+    }
+  }
+}
