+++ frontend/src/pages/product-detail/product-detail.page.jsx
@@ -1,13 +1,12 @@
 import React from "react";
 import "./product-detail.page.scss";
 import {
-  AccentButton,
   BreadcrumbComponent,
-  SearchBarComponent
+  SearchBarComponent,
+  ProductDetail
 } from '../../components';
 import { Row, Col } from 'react-flexbox-grid';
 import { ItemsAPI } from '../../api';
-import NumberFormat from 'react-number-format';
 
 class ProductDetailPage extends React.Component {
   constructor(props) {
@@ -63,47 +62,6 @@ class ProductDetailPage extends React.Component {
       </>
     );
   }
-}
-
-function ProductDetail(props) {
-  const item = props.item;
-  const condition = item.condition === 'new' ? 'Nuevo' : 'Usado';
-  const condition_sold = condition.concat(' - ').concat(item.sold_quantity).concat(' vendidos');
-
-  function comprarAction() {
-    alert('Compraste el producto!!');
-  }
-
-  return (
-    <div className="product">
-      <div className="product-info">
-        <div className="picture">
-          <img
-            src={item.picture}
-            alt={item.title} />
-        </div>
-        <div className="panel">
-          <div className="condition-sold">{condition_sold}</div>
-          <div className="item-title">{item.title}</div>
-          <div className="price">
-            <NumberFormat
-              value={item.price.amount}
-              displayType={'text'}
-              thousandSeparator={'.'}
-              decimalSeparator={','}
-              prefix={'$ '}
-            />
-          </div>
-          <AccentButton onClick={comprarAction}>Comprar</AccentButton>
-        </div>
-      </div>
-      <div className="description">
-        <div className="title">Descripción del producto</div>
-        <div className="text">{item.description}</div>
-      </div>
-    </div>
-  )
-
 }
 
 export { ProductDetailPage };
