+++ frontend/src/pages/product-detail/product-detail.page.scss
@@ -5,62 +5,4 @@
   border-radius: 4px;
   padding: 32px;
   margin-bottom: 120px;
-
-  .product {
-    display: flex;
-    flex-direction: column;
-
-    .product-info {  
-      display: flex;
-      flex-wrap: wrap;
-      margin-bottom: 120px;
-      
-      .picture {
-        width: 680px;
-        max-height: 680px;
-        border-radius: 4px;
-        overflow: hidden;
-        display: flex;
-        flex: 0 1 680px;
-        justify-content: center;
-
-        img {
-          width: 100%;
-          object-fit: contain;
-        }
-      }
-
-      .panel {
-        display: flex;
-        flex: 1 1 200px;
-        flex-direction: column;
-        max-width: 300px;
-
-        .condition-sold {
-          font-size: 14px;
-          margin-bottom: 16px;
-        }
-
-        .item-title {
-          font-size: 24px;
-          margin-bottom: 32px;
-        }
-
-        .price {
-          font-size: 46px;
-          margin-bottom: 32px;
-        }
-      }
-    }
-
-    .description {
-      .title {
-        font-size: 28px;
-        margin-bottom: 32px;
-      }
-      .text {
-        font-size: 16px;
-      }
-    }
-  }
 }
